mulit_language_embedding/requirements.txt
PyPDF2==3.0.1
python-docx==1.1.0
langdetect==1.0.9
sentence-transformers==2.6.0
faiss-cpu==1.8.0
torch==2.2.0
numpy==1.26.4
